feat: fix e2e and add component popover issues
